+++ public/js/app.js
@@ -59861,6 +59861,10 @@ $(function () {
     });
 });
 
+$('.js-records-per-page').change(function () {
+    $(this).closest('form').submit();
+});
+
 // data tables functionality
 $(function () {
     $(document).ready(function () {
@@ -59870,6 +59874,7 @@ $(function () {
         $dataTable.DataTable({
             pageLength: $dataTable.data('page-length') ? $dataTable.data('page-length') : 25,
             responsive: true,
+            searching: pagingFlag,
             paging: pagingFlag,
             bInfo: pagingFlag,
             order: [],